Updating face match algorithm
